Automatically Committed for MakeGamesWithUs

+++ GameOfLife.spritebuilder/Grid.m
@@ -81,9 +81,9 @@ static const int GRID_COLUMNS = 10;
 - (Creature *)creatureForTouchPosition:(CGPoint)touchPosition
 {
     //get the row and column that was touched, return the Creature inside the corresponding cell
-    int row = (int) (touchPosition.y/_cellHeight);
+    float row = (touchPosition.y/_cellHeight);
     int column = (int) (touchPosition.x/_cellWidth);
-    return _gridArray[row][column];
+    return _gridArray[(int)row][column];
 }
 
 @end
